Get the default width and height if the container is not yet rendered (#24)
If the element is rendered in a hidden or collapsed element, or if the DOM hasn’t finished loading at the time of measurement. The clientWidth and clientHeight are zeros. This will result in the wrong size of the render and the position the camera. We use the size from the CSS style in this case or the default values if the CSS style is not set.